// src/app/core/models/enums/rights.ts
export enum Right {
	ROLESLIST = 2,
	ROLESADD = 1,
	ROLESEDIT = 3,
	USERSLIST = 5,
	USERSADD = 4,
	USERSEDIT = 6,
	REQUESTS = 7,
	APPROVEDORGANISATIONS = 8,
	REJECTEORGANISATIONS = 9,
	ORGANIZATIONEDIT = 10,
	ADDNEWCOMMITTEE = 13,
	COMMITTEES = 14,
	COMMITTEEEDIT = 15,
	ADDNEWMEETINGTYPE = 16,
	MEETINGTYPESLIST = 17,
	EDITMEETINGTYPE = 18,
	ADDTIMEZONES = 19,
	TIMEZONESLIST = 20,
	EDITTIMEZONES = 21,
	ADDNEWMEETING = 22,
	MEETINGS = 23,
	EDITMEETING = 24,
	ACTIVE_USER = 25,
	DE_ACTIVE_USER = 26,
	DELETE_ROLE = 27,
	DELETE_USER = 28,
	DELETE_COMMITTEE = 29,
	MEETINGSDashBOARD = 30,
	MANAGEORGANIZATION = 32,

	ADDPROPOSAL = 36,
	PROPOSALS = 37,
	PROPOSALDETAILS = 38,
	DELETE_MEETING_TYPE = 39,
	CALENDAR = 31,

	ADDPARTICIPANT = 33,
	PARTICIPANTS = 34,
	EDITPARTICIPANT = 35,
	DELETE_TIME_ZONE = 40,
	ADMINDASHBOARD = 43,
	SECERTARYDASHBOARD = 41,
	PARTICIPANTDASHBOARD = 42,

	VIEWMEETING = 46,

	MANAGEABSENCE = 48,

	ADDNEWJOBTITLE = 49,
	JOBTITLESLIST = 50,
	EDITJOBTITLE = 51,

	ADDNEWUSERTITLE = 52,
	USERTITLESLIST = 53,
	EDITUSERTITLE = 54,

	ADDNEWNICKNAME = 55,
	NICKNAMESLIST = 56,
	EDITNICKNAME = 57,
	ORGANIZATIONDASHBOARD = 58,
	DELETE_JOB_TITLE = 59,
	DELETE_PARTICIPANT = 60,
	DELETE_USER_TITLE = 61,
	DELETE_NICK_NAME = 62,
	DELETEMEETING = 63,
	MANAGEMOM = 64,
	ADDNEWTASK = 65,
	EDITTASK = 66,
	DELETE_TASK = 67,

	ORGANIZATIONTASKDASHBOARD = 68,
	MYTASKDASHBOARD = 69,
	TASKDETAILS = 70,

	SETTINGSADMIN = 71,
	COMMITTEE_DETAILS = 72,

	ALL_COMMITTEE_TASKS = 74,
	TASKS_STATISTICS = 75,
	CHAT_CONVERSATIONS = 76,
	MEETING_CONVERSATIONS = 77,
	BLOCKED_USERS = 78,
	DELETE_ATTEMPT = 79,
	MOMTEMPLATELIST = 85,
	MOMTEMPLATEADD = 84,
	MOMTEMPLATEEDIT = 86,
	MOMTEMPLATEDELETE = 87,
	LIST_ONLINE_ACCOUNTS = 80,
	ADD_ONLINE_ACCOUNTS = 81,
	EDIT_ONLINE_ACCOUNTS = 82,
	DELETE_ONLINE_ACCOUNTS = 83,
	AGENDATEMPLATELIST = 89,
	AGENDATEMPLATEADD = 88,
	AGENDATEMPLATEEDIT = 90,
	AGENDATEMPLATEDELETE = 91,
	HTMLMOMTEMPLATELIST = 93,
	HTMLMOMTEMPLATEADD = 92,
	HTMLMOMTEMPLATEEDIT = 94,
	HTMLMOMTEMPLATEDELETE = 95,
	REVIEW_ROOM = 96,
	ADD_DOCUMENT = 97,
	EDIT_DOCUMENT = 98,
	DELETE_DOCUMENT = 99,
	REVIEW_DOCUMENT = 100,
	DECISIONS_LIST = 101,
	DECISION_TYPES_LIST = 102,
	ADD_DECISION_TYPE = 103,
	EDIT_DECISION_TYPE = 104,
	DELETE_DECISION_TYPE = 105,
	CIRCULAR_DECISIONS_LIST = 106,
	ADD_CIRCULAR_DECISION = 107,
	EDIT_CIRCULAR_DECISION = 108,
	DELETE_CIRCULAR_DECISION = 109,
	DECISION_DETAILS = 110,
	BOARD_DASHBOARD = 111,
	COMMITTEES_DASHBOARD = 112,
	MEMBER_DASHBOARD = 113,

	ADDFAQSECTION = 116,
	FAQSECTIONLIST = 115,
	EDITFAQSECTION = 117,
	DELETEFAQSECTION = 118,

	ADDFAQ = 120,
	EDITFAQ = 121,
	DELETEFAQ = 122,
	FAQLIST = 119,

	VIDEOS_GUIDE_LIST = 123,
	ADD_VIDEO_GUIDE = 124,
	EDIT_VIDEO_GUIDE = 125,
	DELETE_VIDEO_GUIDE = 126,

	FAQ_TREE = 127,
	VIDEO_GUIDE_HELP = 128,

	STAKEHOLDER_LIST = 129,
	STAKEHOLDER_ADD = 130,
	STAKEHOLDER_EDIT = 131,
	DELETE_STAKEHOLDER = 132,
	DE_ACTIVE_STAKEHOLDER = 133,
	ACTIVE_STAKEHOLDER = 134,

	/* Guest Rights */
	presenting = 137,
	canChat = 138,
	canShowMom = 139,
	canVote = 140,
	canShowVoteResults = 141,
	/* End Guest Rights */

	/* Meeting Page Rights */
	organizationsSection = 142,
	participantsSection = 143,
	attachmentsSection = 144,
	agendasSection = 145,
	votersResultsSection = 146,
	/* End Meeting Page Rights */
	// Approvals
	Approvals = 147,
	addApproval = 148,
	showApproval = 149,
	EDITAPPROVAL = 150,
	DELETEAPPROVAL = 151,

	//Requests
	PermanentCommittees=150,
	TemporaryCommittees=151,
	CommitteeRequests=152,

	COMMITTEEDASHBOARD=154,
	//block user
	BlockUser=153,
	MyCommittees = 158,

	// history
	HISTORY=159
}